pyproject.toml
[tool.poetry]
name = "bdc_api"
version = "0.1.0"
description = "BDC API"
authors = ["Milton Abrunhosa <ml_byteman@yahoo.com.br>"]
license = "CC0 1.0 Universal"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
psycopg2 = "^2.9.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.950"
pylint = "^2.13.7"
autopep8 = "^1.6.0"
types-requests = "^2.27.24"
flake8 = "^4.0.1"
pytest-coverage = "^0.0"
pip-audit = "^2.1.1"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
